Despues de 5 dias de intentos pude configurar Groovy :-P
