+++ pom.xml
@@ -39,11 +39,11 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-          	<groupId>org.codehaus.groovy</groupId>
-          	<artifactId>groovy-all</artifactId>
-          	<version>1.7.0</version>
-        </dependency>
-		
+			<groupId>org.codehaus.groovy</groupId>
+			<artifactId>groovy-all</artifactId>
+			<version>1.7.0</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
@@ -57,39 +57,132 @@
 				</configuration>
 			</plugin>
 			<plugin>
-			      <groupId>org.codehaus.gmaven</groupId>
-			      <artifactId>gmaven-plugin</artifactId>
-			      <version>1.2</version>
-			      <configuration> 
-			        <providerSelection>1.7</providerSelection> 
-			      </configuration> 
-			      <dependencies>
-			          <dependency>
-			            <groupId>org.codehaus.gmaven.runtime</groupId>
-			            <artifactId>gmaven-runtime-1.7</artifactId>
-			            <version>1.2</version>
-			            <exclusions>
-			              <exclusion>
-			                <groupId>org.codehaus.groovy</groupId>
-			                <artifactId>groovy-all</artifactId>
-			              </exclusion>
-			            </exclusions>
-			          </dependency>
-			          <dependency>
-			            <groupId>org.codehaus.groovy</groupId>
-			            <artifactId>groovy-all</artifactId>
-			            <version>1.8.6</version>
-			          </dependency>
-			      </dependencies>
-			      <executions>
-			          <execution>
-			              <goals>
-			                  <goal>compile</goal>
-			                  <goal>testCompile</goal>
-			              </goals>
-			          </execution>
-			      </executions>
-			  </plugin>
+				<groupId>org.codehaus.gmaven</groupId>
+				<artifactId>gmaven-plugin</artifactId>
+				<version>1.2</version>
+				<configuration>
+					<providerSelection>1.7</providerSelection>
+				</configuration>
+				<dependencies>
+					<dependency>
+						<groupId>org.codehaus.gmaven.runtime</groupId>
+						<artifactId>gmaven-runtime-1.7</artifactId>
+						<version>1.2</version>
+						<exclusions>
+							<exclusion>
+								<groupId>org.codehaus.groovy</groupId>
+								<artifactId>groovy-all</artifactId>
+							</exclusion>
+						</exclusions>
+					</dependency>
+					<dependency>
+						<groupId>org.codehaus.groovy</groupId>
+						<artifactId>groovy-all</artifactId>
+						<version>1.8.6</version>
+					</dependency>
+				</dependencies>
+				<executions>
+					<execution>
+						<goals>
+							<goal>compile</goal>
+							<goal>testCompile</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+			<!-- Plugin para agegar groovy como Source Folder -->
+			<plugin>
+				<groupId>org.codehaus.mojo</groupId>
+				<artifactId>build-helper-maven-plugin</artifactId>
+				<version>1.5</version>
+				<executions>
+					<execution>
+						<id>add-source</id>
+						<phase>generate-sources</phase>
+						<goals>
+							<goal>add-source</goal>
+						</goals>
+						<configuration>
+							<sources>
+								<source>src/main/groovy</source>
+							</sources>
+						</configuration>
+					</execution>
+					<execution>
+						<id>add-test-source</id>
+						<phase>generate-test-sources</phase>
+						<goals>
+							<goal>add-test-source</goal>
+						</goals>
+						<configuration>
+							<sources>
+								<source>src/test/groovy</source>
+							</sources>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
 		</plugins>
+
+		<!-- This plugin's configuration is used to store Eclipse m2e settings 
+			only. It has no influence on the Maven build itself. -->
+		<pluginManagement>
+			<plugins>
+				<plugin>
+					<groupId>org.eclipse.m2e</groupId>
+					<artifactId>lifecycle-mapping</artifactId>
+					<version>1.0.0</version>
+					<configuration>
+						<lifecycleMappingMetadata>
+							<pluginExecutions>
+								<pluginExecution>
+									<pluginExecutionFilter>
+										<groupId>
+											org.codehaus.gmaven
+										</groupId>
+										<artifactId>
+											gmaven-plugin
+										</artifactId>
+										<versionRange>
+											[1.2,)
+										</versionRange>
+										<goals>
+											<goal>testCompile</goal>
+										</goals>
+									</pluginExecutionFilter>
+									<action>
+										<execute>
+											<runOnIncremental>false</runOnIncremental>
+										</execute>
+									</action>
+								</pluginExecution>
+								<pluginExecution>
+									<pluginExecutionFilter>
+										<groupId>
+											org.codehaus.gmaven
+										</groupId>
+										<artifactId>
+											gmaven-plugin
+										</artifactId>
+										<versionRange>
+											[1.2,)
+										</versionRange>
+										<goals>
+											<goal>compile</goal>
+										</goals>
+									</pluginExecutionFilter>
+									<action>
+										<execute>
+											<runOnIncremental>false</runOnIncremental>
+										</execute>
+									</action>
+								</pluginExecution>
+							</pluginExecutions>
+						</lifecycleMappingMetadata>
+					</configuration>
+				</plugin>
+			</plugins>
+		</pluginManagement>
+
 	</build>
 </project>
